More test migration

+++ projects2/app/Http/Controllers/ProjectController.php
@@ -63,8 +63,9 @@ class ProjectController extends Controller
             'description' => 'required',
             'courses' => 'required|array',
             'maximum_students' => 'required|integer|min:1',
+            'user_id' => 'required|integer',
         ]);
-        
+
         $project = new Project;
         $project->fill($request->input());
         $project->save();
@@ -124,9 +125,8 @@ class ProjectController extends Controller
             'title' => 'required|max:255',
             'description' => 'required',
             'courses' => 'required|array',
-            'type_id' => 'required|integer',
             'maximum_students' => 'required|integer|min:1',
-            'user_id' => 'required|integer|min:1'
+            'user_id' => 'required|integer',
         ]);
         $project = Project::findOrFail($id);
         if (Gate::denies('edit_this_project', $project)) {
@@ -149,6 +149,9 @@ class ProjectController extends Controller
     public function destroy($id)
     {
         $project = Project::findOrFail($id);
+        if (Gate::denies('edit_this_project', $project)) {
+            abort(403);
+        }
         EventLog::log(Auth::user()->id, "Deleted project {$project->title}");
         $project->delete();
         return redirect()->to('/')->with('success_message', 'Project deleted');
